Upgrading cocoapod to suppress more warnings
