+++ openHABCore/Sources/Model/OpenHABSitemapPage.swift
@@ -112,7 +112,7 @@ extension OpenHABSitemapPage {
 
 extension OpenHABSitemapPage.CodingData {
     public var openHABSitemapPage: OpenHABSitemapPage {
-        let mappedWidgets = widgets?.map { $0.openHABWidget } ?? []
+        let mappedWidgets = widgets?.map(\.openHABWidget) ?? []
         return OpenHABSitemapPage(pageId: pageId ?? "", title: title ?? "", link: link ?? "", leaf: leaf ?? false, widgets: mappedWidgets)
     }
 }
