+++ openHABWatch Extension/Views/Rows/SetpointRow.swift
@@ -37,7 +37,7 @@ struct SetpointRow: View {
 
                 EncircledIconWithAction(systemName:
                     "chevron.down.circle.fill",
-                                        action: self.decreaseValue)
+                    action: self.decreaseValue)
 
                 Spacer()
